extract the modal
