+++ lib/auth/ApplicationAuth.tsx
@@ -1,12 +1,12 @@
 import React, { FunctionComponent, useEffect, useState } from "react";
-import { Button, EmptyState, EmptyStateBody, EmptyStateHeader, EmptyStateIcon, Form, FormGroup, Modal, Spinner, TextInput } from "@patternfly/react-core";
+import { EmptyState, EmptyStateBody, EmptyStateHeader, EmptyStateIcon, Spinner } from "@patternfly/react-core";
 import { ErrorCircleOIcon } from "@patternfly/react-icons";
 import { AuthService, UsernameAndPassword, useAuth } from "./useAuth.ts";
 import { If } from "../common";
+import { BasicAuthModal } from "../modals/BasicAuthModal.tsx";
 
 export interface AuthStateData {
     state: AuthState;
-    creds: UsernameAndPassword;
 }
 
 enum AuthState {
@@ -25,37 +25,13 @@ export type AuthProps = {
  */
 export const ApplicationAuth: FunctionComponent<AuthProps> = (props: AuthProps) => {
     const [authState, setAuthState] = useState<AuthStateData>({
-        state: AuthState.AUTHENTICATING,
-        creds: {
-            username: "",
-            password: ""
-        }
+        state: AuthState.AUTHENTICATING
     });
     const auth: AuthService = useAuth();
 
-    const onUsernameChange = (_event: any, value: string): void => {
-        setAuthState({
-            ...authState,
-            creds: {
-                ...authState.creds,
-                username: value
-            }
-        });
-    };
-
-    const onPasswordChange = (_event: any, value: string): void => {
-        setAuthState({
-            ...authState,
-            creds: {
-                ...authState.creds,
-                password: value
-            }
-        });
-    };
-
-    const basicAuthLogin = (): void => {
+    const basicAuthLogin = (creds: UsernameAndPassword): void => {
         console.info("[ApplicationAuth] Using username and password.");
-        auth.login(authState.creds.username, authState.creds.password);
+        auth.login(creds.username, creds.password);
         setAuthState({
             ...authState,
             state: AuthState.AUTHENTICATED
@@ -99,52 +75,7 @@ export const ApplicationAuth: FunctionComponent<AuthProps> = (props: AuthProps) 
                 </EmptyState>
             </If>
             <If condition={authState.state === AuthState.AUTHENTICATING && auth.isBasicAuthEnabled()}>
-                <Modal
-                    title="Login"
-                    variant="small"
-                    isOpen={true}
-                    header={<a href="#" />}
-                    showClose={false}
-                    className="please-wait pf-m-redhat-font"
-                    aria-label="please-wait-modal"
-                    style={{ marginTop: "-15px" }}
-                    actions={[
-                        <Button key="login" variant="primary" data-testid="modal-btn-login" onClick={basicAuthLogin} isDisabled={authState.creds.username === "" || authState.creds.password === ""}>Login</Button>
-                    ]}
-                >
-                    <Form>
-                        <FormGroup
-                            label="Username"
-                            fieldId="form-username"
-                        >
-                            <TextInput
-                                isRequired={true}
-                                type="text"
-                                id="form-username"
-                                data-testid="basic-auth-login-modal-username"
-                                name="form-username"
-                                value={authState.creds.username}
-                                placeholder="Username"
-                                onChange={onUsernameChange}
-                            />
-                        </FormGroup>
-                        <FormGroup
-                            label="Password"
-                            fieldId="form-password"
-                        >
-                            <TextInput
-                                isRequired={true}
-                                type="text"
-                                id="form-password"
-                                data-testid="basic-auth-login-modal-password"
-                                name="form-password"
-                                value={authState.creds.password}
-                                placeholder="Password"
-                                onChange={onPasswordChange}
-                            />
-                        </FormGroup>
-                    </Form>
-                </Modal>
+                <BasicAuthModal onLogin={basicAuthLogin}></BasicAuthModal>
             </If>
             <If condition={authState.state === AuthState.AUTHENTICATION_FAILED}>
                 <EmptyState>
